CortePerfecto/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.3.2",
    "plotly>=6.3.0",
    "psycopg2-binary>=2.9.10",
    "reportlab>=4.4.4",
    "streamlit>=1.50.0",
    "xlsxwriter>=3.2.9",
]
